Create a separate IBMJDKtoOracleJDKTest
