+++ src/test/java/org/openrewrite/java/migrate/IBMSemeruTest.java
@@ -42,7 +42,7 @@ class IBMSemeruTest implements RewriteTest {
           java(
             """
               import com.sun.net.ssl.internal.www.protocol.https.*;  //do NOT flag this
-                            
+              
               class Foo{
                 void bar() {
                     com.sun.net.ssl.internal.www.protocol.https.Handler handler_1 =           //flag
@@ -53,12 +53,12 @@ class IBMSemeruTest implements RewriteTest {
                     if (handler_1 instanceof com.sun.net.ssl.internal.www.protocol.https.Handler){ //flag
                         //do nothing
                     }
-                            
+              
                     if (handler_1 instanceof Handler){ //flag
                         //do nothing
                     }
                 }
-                            
+              
                 public static com.sun.net.ssl.internal.www.protocol.https.Handler testMethod(Handler handler){ //flag (2)
                     return handler;
                 }
@@ -67,7 +67,7 @@ class IBMSemeruTest implements RewriteTest {
             """
               import com.ibm.net.ssl.www2.protocol.https.Handler;
               import com.sun.net.ssl.internal.www.protocol.https.*;  //do NOT flag this
-                            
+              
               class Foo{
                 void bar() {
                     com.ibm.net.ssl.www2.protocol.https.Handler handler_1 =           //flag
@@ -78,12 +78,12 @@ class IBMSemeruTest implements RewriteTest {
                     if (handler_1 instanceof com.ibm.net.ssl.www2.protocol.https.Handler){ //flag
                         //do nothing
                     }
-                            
+              
                     if (handler_1 instanceof Handler){ //flag
                         //do nothing
                     }
                 }
-                            
+              
                 public static com.ibm.net.ssl.www2.protocol.https.Handler testMethod(Handler handler){ //flag (2)
                     return handler;
                 }
@@ -101,7 +101,7 @@ class IBMSemeruTest implements RewriteTest {
             """
               class Foo{
                 private String flagMe = "com.sun.net.ssl.internal.www.protocol"; //flag this
-                            
+              
                 void bar() {
                     System.setProperty("java.protocol.handler.pkgs", "com.sun.net.ssl.internal.www.protocol"); //flag this
                     String s1 = "com.sun.net.ssl";                              //DO NOT FLAG
@@ -116,7 +116,7 @@ class IBMSemeruTest implements RewriteTest {
             """
               class Foo{
                 private String flagMe = "com.ibm.net.ssl.www2.protocol"; //flag this
-                            
+              
                 void bar() {
                     System.setProperty("java.protocol.handler.pkgs", "com.ibm.net.ssl.www2.protocol"); //flag this
                     String s1 = "com.sun.net.ssl";                              //DO NOT FLAG
@@ -139,24 +139,24 @@ class IBMSemeruTest implements RewriteTest {
           java(
             """
               import com.sun.net.ssl.internal.ssl.*;  // do NOT flag, handled by other rule
-                            
+              
               class TestClass_2{
                 void bar() {
                     Provider provider_4 = new Provider();  // flag (2)
                 }
-                            
+              
                 private void fdsa( Provider p1 ){} // flag
               }
               """,
             """
               import com.ibm.jsse2.IBMJSSEProvider2;
               import com.sun.net.ssl.internal.ssl.*;  // do NOT flag, handled by other rule
-                            
+              
               class TestClass_2{
                 void bar() {
                     IBMJSSEProvider2 provider_4 = new IBMJSSEProvider2();  // flag (2)
                 }
-                            
+              
                 private void fdsa( IBMJSSEProvider2 p1 ){} // flag
               }
               """
@@ -183,7 +183,7 @@ class IBMSemeruTest implements RewriteTest {
                 import com.sun.net.ssl.TrustManagerFactorySpi;
                 import com.sun.net.ssl.X509KeyManager;
                 import com.sun.net.ssl.X509TrustManager;
-                               
+                
                 class TestFullyQualifiedPackage {
                     com.sun.net.ssl.HostnameVerifier hv;
                     com.sun.net.ssl.HttpsURLConnection huc;
@@ -198,7 +198,7 @@ class IBMSemeruTest implements RewriteTest {
                     com.sun.net.ssl.TrustManagerFactorySpi tmfs;
                     com.sun.net.ssl.X509KeyManager x509km;
                     com.sun.net.ssl.X509TrustManager xtm;
-                                
+                
                     HostnameVerifier hv2;
                     HttpsURLConnection huc2;
                     KeyManager km2;
@@ -228,7 +228,7 @@ class IBMSemeruTest implements RewriteTest {
                 import javax.net.ssl.TrustManagerFactorySpi;
                 import javax.net.ssl.X509KeyManager;
                 import javax.net.ssl.X509TrustManager;
-                              
+                
                 class TestFullyQualifiedPackage {
                     javax.net.ssl.HostnameVerifier hv;
                     javax.net.ssl.HttpsURLConnection huc;
@@ -243,7 +243,7 @@ class IBMSemeruTest implements RewriteTest {
                     javax.net.ssl.TrustManagerFactorySpi tmfs;
                     javax.net.ssl.X509KeyManager x509km;
                     javax.net.ssl.X509TrustManager xtm;
-                                
+                
                     HostnameVerifier hv2;
                     HttpsURLConnection huc2;
                     KeyManager km2;
@@ -270,7 +270,7 @@ class IBMSemeruTest implements RewriteTest {
             //language=java
             java("""
                 import com.sun.net.ssl.HostnameVerifier;
-                               
+                
                 class TestHostnameVerifier implements HostnameVerifier {
                     public boolean verify(String arg0, String arg1) {
                         return false;
@@ -279,7 +279,7 @@ class IBMSemeruTest implements RewriteTest {
                 """,
               """
                 import javax.net.ssl.HostnameVerifier;
-                               
+                
                 class TestHostnameVerifier implements HostnameVerifier {
                     public boolean verify(String arg0, String arg1) {
                         return false;
